build event on the stack then record
Build the event on the stack before recording it to the log,
so that we can later make sure the calls to the clocks are
made with maximum use of the cache.

This requires us to not allocate the clock together with the events,
so that we do not confuse memory that is often read from memory
that is often written.