improve cache locality by having clock close in read arena
Clock lookups were slow because we had the clock in the event arena,
in close quarters with the writable data. Pulling it out in the read
section reduces overhead when occupying multiple cores.

A further optimization would be to not having to read anything. This
would be accomplished by storing opaque timestamps and doing the
conversion to microseconds at reporting time only.